add more 2015 days

+++ java/src/main/java/fr/akshell/aoc/y2015/Day5.java
@@ -6,7 +6,7 @@ import java.util.Arrays;
 import java.util.function.Function;
 import java.util.regex.Pattern;
 
-public class Day5 extends BaseDay<Long> {
+public class Day5 extends BaseDay<Integer> {
     
     // rule 1 V1: at least 3 vowels
     private static final Pattern RE_RULE1_V1 = Pattern.compile("[aeiou].*[aeiou].*[aeiou]");
@@ -29,18 +29,18 @@ public class Day5 extends BaseDay<Long> {
             string -> RE_RULE1_V2.matcher(string).find()
                     && RE_RULE2_V2.matcher(string).find();
 
-    public long countNiceStrings(Function<String, Boolean> niceStringTest, String input) {
+    public int countNiceStrings(Function<String, Boolean> niceStringTest, String input) {
             return Arrays
                     .stream(input.split("\n"))
-                    .mapToLong(line -> niceStringTest.apply(line) ? 1 : 0)
+                    .mapToInt(line -> niceStringTest.apply(line) ? 1 : 0)
                     .sum();
     }
 
-    public Long part1(String input) {
+    public Integer part1(String input) {
         return countNiceStrings(NICE_STRING_V1, input);
     }
 
-    public Long part2(String input) {
+    public Integer part2(String input) {
         return countNiceStrings(NICE_STRING_V2, input);
     }
 }
